apps/accelerate/cloud_surfer/pyproject.toml
[tool.poetry]
name = "cloud-surfer"
version = "0.1.0"
description = "Cloud Surfer"
authors = [
    "Michel Zanotti <m.zanotti@nebuly.ai>",
    "Diego Fiori <d.fiori@nebuly.ai>",
]
readme = "README.md"
packages = [{ include = "surfer" }]

[tool.poetry.scripts]
surfer = "surfer.cli.main:main"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.7.0"
ray = {extras = ["default"], version = "^2.3.0"}
speedster = "^0.2.1"
rich = "^13.3.1"
pyyaml = "^6.0"
pydantic = "^1.10.5"
random-word = "^1.0.11"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
coverage = "^7.2.1"


[tool.poetry.group.gcp.dependencies]
google-cloud-storage = "^2.7.0"

[tool.pytest.ini_options]
addopts = [
  "--strict-config",
  "--strict-markers",
]
minversion = "7.0"
testpaths = [
    "tests",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
